Suggestion to structure changes
