+++ src/components/Elements/Button/Button.Props.tsx
@@ -1,0 +1,24 @@
+import * as React from 'react';
+
+const variants = {
+  primary: 'bg-blue-600 text-white',
+  inverse: 'bg-white text-blue-600',
+  danger: 'bg-red-600 text-white',
+};
+
+const sizes = {
+  sm: 'py-2 px-4 text-sm',
+  md: 'py-2 px-6 text-md',
+  lg: 'py-3 px-8 text-lg',
+};
+
+type IconProps =
+  | { startIcon: React.ReactElement; endIcon?: never }
+  | { endIcon: React.ReactElement; startIcon?: never }
+  | { endIcon?: undefined; startIcon?: undefined };
+
+export type ButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement> & {
+  variant?: keyof typeof variants;
+  size?: keyof typeof sizes;
+  isLoading?: boolean;
+} & IconProps;
