+++ src/components/Elements/Button/Button.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-
 import { Spinner } from '@/components/Elements';
+import { ButtonProps } from './Button.Props';
 import clsx from 'clsx';
 
 const variants = {
@@ -14,17 +14,6 @@ const sizes = {
   md: 'py-2 px-6 text-md',
   lg: 'py-3 px-8 text-lg',
 };
-
-type IconProps =
-  | { startIcon: React.ReactElement; endIcon?: never }
-  | { endIcon: React.ReactElement; startIcon?: never }
-  | { endIcon?: undefined; startIcon?: undefined };
-
-export type ButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement> & {
-  variant?: keyof typeof variants;
-  size?: keyof typeof sizes;
-  isLoading?: boolean;
-} & IconProps;
 
 export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   (
